deps/v8/test/inspector/debugger/set-breakpoint-url-regex-expected.txt

Tests for calling setBreakpoint with urlRegex

Running test: testSetBreakpointByUrlRegex
[
    [0] : {
        columnNumber : 2
        lineNumber : 2
        scriptId : <scriptId>
    }
]
Successfully paused during eval of: 'test()'
Successfully completed eval of: 'test()'

Running test: testSetBreakpointByUrlWithConditions
[
    [0] : {
        columnNumber : 2
        lineNumber : 2
        scriptId : <scriptId>
    }
]
Successfully completed eval of: 'test()'
[
    [0] : {
        columnNumber : 2
        lineNumber : 2
        scriptId : <scriptId>
    }
]
Successfully paused during eval of: 'test()'
